+++ packages/apps/dev-wallet/src/pages/home/home-page.tsx
@@ -60,6 +60,9 @@ export function HomePage() {
           </ul>
         ) : null}
       </Box>
+      <Box>
+        <Link to="/sig-builder">Sig Builder</Link>
+      </Box>
     </main>
   );
 }