+++ packages/apps/proof-of-us/src/components/DetailView/DetailView.tsx
@@ -1,4 +1,5 @@
 import { Button } from '@/components/Button/Button';
+import { MessageBlock } from '@/components/MessageBlock/MessageBlock';
 import { useAvatar } from '@/hooks/avatar';
 import { useProofOfUs } from '@/hooks/proofOfUs';
 import { isAlreadySigning } from '@/utils/isAlreadySigning';
@@ -16,7 +17,7 @@ import { ImagePositions } from '../ImagePositions/ImagePositions';
 import { ScreenHeight } from '../ScreenHeight/ScreenHeight';
 import { TextField } from '../TextField/TextField';
 import { TitleHeader } from '../TitleHeader/TitleHeader';
-import { imageWrapper, titleErrorClass } from './style.css';
+import { imageWrapper } from './style.css';
 
 interface IProps {
   next: () => void;
@@ -115,10 +116,10 @@ export const DetailView: FC<IProps> = ({ next, prev }) => {
       )}
 
       <Stack flex={1} />
+      {titleError && <MessageBlock variant="error">{titleError}</MessageBlock>}
       <Button variant="primary" onPress={handleShare}>
         Share <MonoQrCodeScanner />
       </Button>
-      {titleError && <div className={titleErrorClass}>{titleError}</div>}
     </ScreenHeight>
   );
 };