+++ packages/apps/proof-of-us/src/components/ShareView/ShareView.tsx
@@ -7,7 +7,7 @@ import { isAlreadySigning, isSignedOnce } from '@/utils/isAlreadySigning';
 import {
   MonoArrowBack,
   MonoArrowDownward,
-  MonoCheck,
+  MonoCheckCircle,
 } from '@kadena/react-icons';
 import { Stack } from '@kadena/react-ui';
 import Link from 'next/link';
@@ -21,7 +21,7 @@ import { TitleHeader } from '../TitleHeader/TitleHeader';
 
 import { useAccount } from '@/hooks/account';
 import { ScreenHeight } from '../ScreenHeight/ScreenHeight';
-import { qrClass } from './style.css';
+import { copyClass, qrClass } from './style.css';
 
 interface IProps {
   next: () => void;
@@ -70,6 +70,16 @@ export const ShareView: FC<IProps> = ({ prev, status }) => {
     //updateStatus({ proofOfUsId: proofOfUs.proofOfUsId, status: 4 });
   }, []);
 
+  useEffect(() => {
+    if (!isCopied) return;
+
+    const timer = setTimeout(() => {
+      setIsCopied(false);
+    }, 3000);
+
+    return () => clearTimeout(timer);
+  }, [isCopied]);
+
   if (!proofOfUs || !account || !isMounted) return;
 
   const handleCopy = () => {
@@ -94,6 +104,15 @@ export const ShareView: FC<IProps> = ({ prev, status }) => {
               </>
             )}
             label="Share"
+            Append={() => (
+              <>
+                {isCopied ? (
+                  <Stack>
+                    Copied <MonoCheckCircle className={copyClass} />
+                  </Stack>
+                ) : null}
+              </>
+            )}
           />
 
           {!isAlreadySigning(proofOfUs.signees) ? (
@@ -105,7 +124,7 @@ export const ShareView: FC<IProps> = ({ prev, status }) => {
               >
                 <QRCode
                   ecLevel="H"
-                  size={qrContainerRef.current?.offsetWidth || 300}
+                  size={300}
                   ref={qrRef}
                   value={`${getReturnHostUrl()}/scan/${proofOfUs.proofOfUsId}`}
                   removeQrCodeBehindLogo={true}
@@ -116,11 +135,6 @@ export const ShareView: FC<IProps> = ({ prev, status }) => {
                 />
               </div>
               <Button onPress={handleCopy}>Click to copy link</Button>
-              {isCopied ? (
-                <Stack>
-                  Copied! <MonoCheck />
-                </Stack>
-              ) : null}
               <ListSignees />
             </>
           ) : (