+++ packages/tools/kadena-cli/src/tx/commands/txList.ts
@@ -3,7 +3,7 @@ import { createCommand } from '../../utils/createCommand.js';
 import { log } from '../../utils/logger.js';
 import { txOptions } from '../txOptions.js';
 import { printTx } from '../utils/txDisplayHelper.js';
-import { getTransactions } from '../utils/txHelpers.js';
+import { getAllTransactions } from '../utils/txHelpers.js';
 
 export const createTxListCommand: (program: Command, version: string) => void =
   createCommand(
@@ -14,15 +14,11 @@ export const createTxListCommand: (program: Command, version: string) => void =
       log.debug('list-tx:action');
 
       const { directory } = await option.directory();
-      const transactions: string[] = await getTransactions(
-        false,
-        directory,
-        true,
-      );
+      const transactions = await getAllTransactions(directory);
 
       transactions.sort((a, b) => {
-        const aIsSigned = a.includes('-signed') ? 1 : -1;
-        const bIsSigned = b.includes('-signed') ? 1 : -1;
+        const aIsSigned = a.signed === true ? 1 : -1;
+        const bIsSigned = b.signed === true ? 1 : -1;
         return bIsSigned - aIsSigned;
       });