+++ packages/libs/react-ui/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @kadena/react-ui
 
+## 0.8.3
+
+### Patch Changes
+
+- 25de564cd: Add expand animation to accordion component
+
+## 0.8.2
+
+### Patch Changes
+
+- 550e6d9e4: Extended clickable area for the numberField buttons
+
 ## 0.8.1
 
 ### Patch Changes