+++ packages/tools/cookbook/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @kadena/cookbook
 
+## 0.0.14
+
+### Patch Changes
+
+- Updated dependencies [d67b52906]
+  - @kadena/client@1.10.1
+
+## 0.0.13
+
+### Patch Changes
+
+- Updated dependencies [016b9dbfc]
+  - @kadena/client@1.10.0
+
+## 0.0.12
+
+### Patch Changes
+
+- Updated dependencies [b53c2600c]
+- Updated dependencies [6ddf094d8]
+- Updated dependencies [6ddf094d8]
+  - @kadena/client@1.9.0
+  - @kadena/chainweb-node-client@0.5.3
+  - @kadena/cryptography-utils@0.4.2
+
 ## 0.0.11
 
 ### Patch Changes