+++ packages/tools/pactjs-cli/CHANGELOG.md
@@ -1,5 +1,32 @@
 # Change Log - @kadena/pactjs-cli
 
+## 1.10.1
+
+### Patch Changes
+
+- Updated dependencies [d67b52906]
+  - @kadena/pactjs-generator@1.10.1
+  - @kadena/client@1.10.1
+
+## 1.10.0
+
+### Patch Changes
+
+- Updated dependencies [016b9dbfc]
+  - @kadena/client@1.10.0
+  - @kadena/pactjs-generator@1.10.0
+
+## 1.9.0
+
+### Patch Changes
+
+- Updated dependencies [b53c2600c]
+- Updated dependencies [6ddf094d8]
+- Updated dependencies [6ddf094d8]
+- Updated dependencies [be9129381]
+  - @kadena/client@1.9.0
+  - @kadena/pactjs-generator@1.9.0
+
 ## 1.8.1
 
 ### Patch Changes